+++ packages/core-js/features/string/index.js
@@ -1,5 +1,6 @@
 var parent = require('../../es/string');
 require('../../modules/esnext.string.at');
+require('../../modules/esnext.string.code-points');
 // TODO: remove from `core-js@4`
 require('../../modules/esnext.string.match-all');
 require('../../modules/esnext.string.replace-all');