+++ packages/core-js/modules/esnext.iterator.constructor.js
@@ -4,6 +4,7 @@ var $ = require('../internals/export');
 var global = require('../internals/global');
 var anInstance = require('../internals/an-instance');
 var createNonEnumerableProperty = require('../internals/create-non-enumerable-property');
+var fails = require('../internals/fails');
 var has = require('../internals/has');
 var wellKnownSymbol = require('../internals/well-known-symbol');
 var IteratorPrototype = require('../internals/iterators-core').IteratorPrototype;
@@ -15,7 +16,11 @@ var TO_STRING_TAG = wellKnownSymbol('toStringTag');
 var NativeIterator = global.Iterator;
 
 // FF56- have non-standard global helper `Iterator`
-var FORCED = IS_PURE || typeof NativeIterator != 'function' || NativeIterator.prototype !== IteratorPrototype;
+var FORCED = IS_PURE
+  || typeof NativeIterator != 'function'
+  || NativeIterator.prototype !== IteratorPrototype
+  // FF44- non-standard `Iterator` passes previous tests
+  || !fails(function () { NativeIterator({}); });
 
 var IteratorConstructor = function Iterator() {
   anInstance(this, IteratorConstructor);
@@ -32,7 +37,7 @@ if (!has(IteratorPrototype, TO_STRING_TAG)) {
   createNonEnumerableProperty(IteratorPrototype, TO_STRING_TAG, 'Iterator');
 }
 
-if (!has(IteratorPrototype, 'constructor') || IteratorPrototype.constructor === Object) {
+if (FORCED || !has(IteratorPrototype, 'constructor') || IteratorPrototype.constructor === Object) {
   createNonEnumerableProperty(IteratorPrototype, 'constructor', IteratorConstructor);
 }