+++ packages/core-js-compat/src/data.js
@@ -717,7 +717,9 @@ const data = {
     // V8 6.6 has a serious bug
     chrome: '67', // '63',
     firefox: '69',
-    safari: '12.0',
+    // Previous versions are non-generic
+    // https://bugs.webkit.org/show_bug.cgi?id=200788
+    safari: '13.0.3', // '12.0',
   },
   'es.reflect.apply': {
     edge: '15',