+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Changelog
+##### Unreleased
+- Fixed missed `esnext.string.code-points` in `core-js/features/string` entry point
+
 ##### 3.4.3 - 2019.11.26
 - Fixed missed `es.json.stringify` and some modules from iteration helpers proposal in some entry points
 - Added a workaround of `String#{ endsWith, startsWith }` MDN polyfills bugs, [#702](https://github.com/zloirock/core-js/issues/702)