+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Changelog
-##### Unreleased
+##### 3.4.4 - 2019.11.27
 - Added feature detection for Safari [non-generic `Promise#finally` bug](https://bugs.webkit.org/show_bug.cgi?id=200829) **(critical for `core-js-pure`)**
 - Fixed missed `esnext.string.code-points` in `core-js/features/string` entry point
 - Updated `Iterator` proposal feature detection for the case of non-standard `Iterator` in FF44-